requirements.txt
amqp==5.3.1
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.10.0
asttokens==3.0.0
attrs==25.4.0
azure-ai-agents==1.2.0b6
azure-ai-projects==1.0.0
azure-cognitiveservices-speech==1.45.0
azure-common==1.1.28
azure-core==1.35.0
azure-cosmos==4.14.1
azure-identity==1.25.1
azure-search-documents==11.6.0
azure-storage-blob==12.27.1
beautifulsoup4==4.13.4
billiard==4.2.2
build==1.3.0
celery==5.5.3
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cryptography==46.0.3
distro==1.9.0
docutils==0.22.2
et_xmlfile==2.0.0
executing==2.2.1
fastapi==0.120.1
googleapis-common-protos==1.71.0
greenlet==3.2.4
grpcio==1.76.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
id==1.5.0
idna==3.11
importlib_metadata==8.7.0
isodate==0.7.2
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.3.0
jeepney==0.9.0
Jinja2==3.1.6
jiter==0.10.0
jiwer==4.0.0
joblib==1.5.2
json5==0.12.1
json_repair==0.25.2
jsonpatch==1.33
jsonpickle==4.1.1
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
keyring==25.6.0
kombu==5.5.4
langchain==0.3.27
langchain-core==0.3.76
langchain-openai==0.3.33
langchain-text-splitters==0.3.11
langdetect==1.0.9
langsmith==0.4.31
lark==1.2.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
more-itertools==10.8.0
msal==1.34.0
msal-extensions==1.3.1
nh3==0.3.1
numpy==2.3.4
oauthlib==3.3.1
openai==1.109.1
openpyxl==3.1.5
opentelemetry-api==1.36.0
opentelemetry-exporter-otlp==1.36.0
opentelemetry-exporter-otlp-proto-common==1.36.0
opentelemetry-exporter-otlp-proto-grpc==1.36.0
opentelemetry-exporter-otlp-proto-http==1.36.0
opentelemetry-proto==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
orjson==3.11.4
packaging==25.0
pandas==2.3.3
pandocfilters==1.5.1
prompt_toolkit==3.0.52
protobuf==6.33.0
pure_eval==0.2.3
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
pydub==0.25.1
Pygments==2.19.2
PyJWT==2.10.1
pyodbc==5.3.0
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pywin32-ctypes==0.2.3
PyYAML==6.0.3
RapidFuzz==3.14.1
readme_renderer==44.0
referencing==0.37.0
regex==2025.10.23
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986==2.0.0
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.2.0
rpds-py==0.28.0
SecretStorage==3.4.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.43
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.47.3
tenacity==9.1.2
tiktoken==0.12.0
tqdm==4.67.1
twine==6.2.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
vine==5.1.0
wcwidth==0.2.14
whatsapp_utilsFAPI==0.20.0
zipp==3.23.0
zstandard==0.25.0